+++ mqtt/mqtt-bridge/src/persist/waking_state/mod.rs
@@ -72,8 +72,7 @@ mod tests {
         // insert a bunch of elements
         let num_elements = 10_usize;
         for i in 0..num_elements {
-            #[allow(clippy::cast_possible_truncation)]
-            let key = Key { offset: i as u32 };
+            let key = Key { offset: i as u64 };
             let publication = Publication {
                 topic_name: i.to_string(),
                 qos: QoS::ExactlyOnce,
@@ -88,7 +87,7 @@ mod tests {
         let mut elements = state.batch(num_elements).unwrap();
         for count in 0..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let num_elements = count as u32;
+            let num_elements = count as u64;
             assert_eq!(elements.pop_front().unwrap().0.offset, num_elements)
         }
     }
@@ -98,8 +97,7 @@ mod tests {
         // insert a bunch of elements
         let num_elements = 10_usize;
         for i in 0..num_elements {
-            #[allow(clippy::cast_possible_truncation)]
-            let key = Key { offset: i as u32 };
+            let key = Key { offset: i as u64 };
             let publication = Publication {
                 topic_name: i.to_string(),
                 qos: QoS::ExactlyOnce,
@@ -125,8 +123,7 @@ mod tests {
 
         // check that the ordering is maintained
         for count in 2..num_elements {
-            #[allow(clippy::cast_possible_truncation)]
-            let num_elements = count as u32;
+            let num_elements = count as u64;
             let extracted_offset = loader.try_next().await.unwrap().unwrap().0.offset;
             assert_eq!(extracted_offset, num_elements)
         }
